backend/requirements.txt
fastapi==0.115.2
pydantic==2.9.2
boto3==1.35.39
pandas==2.2.3
python-dotenv==1.0.1